+++ test/e2e/all_in_one_test.go
@@ -256,6 +256,23 @@ func TestDeployAllInOneEnterprisePostgres(t *testing.T) {
 	verifyEnterpriseWithPostgres(ctx, t, env, adminPassword)
 }
 
+func TestDeployAllInOnePostgresGatewayDiscovery(t *testing.T) {
+	t.Parallel()
+
+	const manifestFilePath = "manifests/all-in-one-postgres-multiple-gateways.yaml"
+
+	t.Logf("configuring %s manifest test", manifestFilePath)
+	ctx, env := setupE2ETest(t)
+
+	t.Log("deploying kong components")
+	deployments := ManifestDeploy{Path: manifestFilePath}.Run(ctx, t, env)
+
+	t.Log("running ingress tests to verify all-in-one deployed ingress controller and proxy are functional")
+	deployIngressWithEchoBackends(ctx, t, env, numberOfEchoBackends)
+	verifyIngressWithEchoBackends(ctx, t, env, numberOfEchoBackends)
+	ensureAllProxyReplicasAreConfigured(ctx, t, env, deployments.ProxyNN)
+}
+
 func TestDeployAllInOneDBLESS(t *testing.T) {
 	t.Parallel()